+++ lib/src/get_it_future_builder2.dart
@@ -46,6 +46,11 @@ class GetItFutureBuilder2<T1 extends Object, T2 extends Object>
     return FutureBuilder(
       builder: (context, snapshot) {
         if (snapshot.hasData) {
+          // ignore: only_throw_errors
+          throw snapshot.error!;
+        }
+
+        if (snapshot.hasData) {
           return _ready(
             context,
             snapshot.data![0] as T1,